Fixed the bug of double dots
