+++ Wire-iOS/Resources/Classy/stylesheet.cas
@@ -523,13 +523,6 @@ ConversationRootViewController {
     }
 }
 
-NewDevicesDot {
-    backgroundColor: $color-accent-current;
-    layer: @{
-        cornerRadius: 4.0;
-    }
-}
-
 ConversationListAccessoryView {
     textLabel: @{
         font: $font-medium-semibold;
