+++ Wire-iOS/Sources/UserInterface/ConversationList/ConversationListViewController+TopBar.swift
@@ -26,7 +26,8 @@ extension ConversationListViewController {
         guard let currentAccount = SessionManager.shared?.accountManager.selectedAccount else {
             fatal("No account available")
         }
-        let currentAccountView = AccountViewFactory.viewFor(account: currentAccount)
+        let currentAccountView = AccountViewFactory.viewFor(account: currentAccount,
+                                                            user: ZMUser.selfUser(inUserSession: ZMUserSession.shared()!))
         currentAccountView.invertUnreadMessagesCount = true
         return currentAccountView
     }
@@ -45,19 +46,8 @@ extension ConversationListViewController {
         profileAccountView.accessibilityHint = "self.voiceover.hint".localized
         
         if let user = ZMUser.selfUser() {
-            let imageView = profileAccountView.imageViewContainer
-            
-            let newDevicesDot = NewDevicesDot(user: user)
-            profileAccountView.addSubview(newDevicesDot)
-            
             if user.clientsRequiringUserAttention.count > 0 {
                 profileAccountView.accessibilityLabel = "self.new-device.voiceover.label".localized
-            }
-            
-            constrain(newDevicesDot, imageView) { newDevicesDot, imageView in
-                newDevicesDot.center.layout(around: imageView.center, at: CGFloat(Double.pi / 4) * 3, diameter: 16)
-                newDevicesDot.width == 8
-                newDevicesDot.height == newDevicesDot.width
             }
         }
