+++ Wire-iOS/Sources/UserInterface/SelfProfile/AccountView.swift
@@ -59,52 +59,6 @@ class ShapeView: LayerHostView<CAShapeLayer> {
     }
 }
 
-class DotView: UIView {
-    fileprivate let circleView = ShapeView()
-    fileprivate let centerView = ShapeView()
-    private var userObserver: NSObjectProtocol!
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        circleView.pathGenerator = {
-            return UIBezierPath(ovalIn: CGRect(origin: .zero, size: $0))
-        }
-        circleView.hostedLayer.lineWidth = 0
-        circleView.hostedLayer.fillColor = UIColor.white.cgColor
-        
-        centerView.pathGenerator = {
-            return UIBezierPath(ovalIn: CGRect(origin: .zero, size: $0))
-        }
-        centerView.hostedLayer.fillColor = UIColor.accent().cgColor
-        
-        addSubview(circleView)
-        addSubview(centerView)
-        constrain(self, circleView, centerView) { selfView, backingView, centerView in
-            backingView.edges == selfView.edges
-            centerView.edges == inset(selfView.edges, 1, 1, 1, 1)
-        }
-        
-        if let userSession = ZMUserSession.shared(),
-            let selfUser = ZMUser.selfUser() {
-            userObserver = UserChangeInfo.add(observer: self, for: selfUser, userSession: userSession)
-        }
-    }
-    
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
-extension DotView: ZMUserObserver {
-    func userDidChange(_ changeInfo: UserChangeInfo) {
-        guard changeInfo.accentColorValueChanged && changeInfo.user.isSelfUser else {
-            return
-        }
-        
-        centerView.hostedLayer.fillColor = UIColor.accent().cgColor
-    }
-}
-
 public protocol AccountViewType {
     var collapsed: Bool { get set }
     var hasUnreadMessages: Bool { get }
@@ -114,8 +68,9 @@ public protocol AccountViewType {
 }
 
 public final class AccountViewFactory {
-    public static func viewFor(account: Account) -> BaseAccountView {
-        return account.teamName == nil ? PersonalAccountView(account: account) : TeamAccountView(account: account)
+    public static func viewFor(account: Account, user: ZMUser? = nil) -> BaseAccountView {
+        return account.teamName == nil ? PersonalAccountView(account: account, user: user)
+            : TeamAccountView(account: account, user: user)
     }
 }
 
@@ -124,9 +79,10 @@ public class BaseAccountView: UIView, AccountViewType {
     
     internal let imageViewContainer = UIView()
     fileprivate let outlineView = UIView()
-    fileprivate let dotView = DotView()
+    fileprivate let dotView : DotView
     fileprivate let selectionView = ShapeView()
     fileprivate var unreadCountToken : Any?
+    public let account: Account
     
     private var selfUserObserver: NSObjectProtocol!
 
@@ -142,8 +98,6 @@ public class BaseAccountView: UIView, AccountViewType {
         }
     }
     
-    /// Set this to false if the account view should show the unread message dot for messages belonging only to this
-    /// account. Set to true if the dot should show for any unread message belonging to all other accounts.
     public var invertUnreadMessagesCount = false
     
     public var hasUnreadMessages: Bool {
@@ -154,12 +108,11 @@ public class BaseAccountView: UIView, AccountViewType {
         }
     }
     
-    public let account: Account
     
     func updateAppearance() {
-        selectionView.isHidden = !selected || collapsed
-        dotView.isHidden = !hasUnreadMessages || collapsed
         
+        selectionView.isHidden = !selected || collapsed
+        dotView.hasUnreadMessages = hasUnreadMessages
         selectionView.hostedLayer.strokeColor = UIColor.accent().cgColor
     }
     
@@ -170,8 +123,11 @@ public class BaseAccountView: UIView, AccountViewType {
                 (self.hasUnreadMessages ? (" " + "conversation_list.header.self_team.accessibility_value.has_new_messages".localized) : "")
     }
     
-    init(account: Account) {
+    init(account: Account, user: ZMUser? = nil) {
         self.account = account
+        
+        dotView = DotView(frame: .zero, hasUnreadMessages: account.unreadConversationCount > 0, user: user)
+        
         super.init(frame: .zero)
         
         if let userSession = SessionManager.shared?.activeUserSession {
@@ -272,8 +228,8 @@ public final class PersonalAccountView: BaseAccountView {
         }
     }
     
-    override init(account: Account) {
-        super.init(account: account)
+    override init(account: Account, user: ZMUser? = nil) {
+        super.init(account: account, user: user)
         
         self.isAccessibilityElement = true
         self.accessibilityTraits = UIAccessibilityTraitButton
@@ -420,11 +376,11 @@ public final class TeamImageView: UIImageView {
     private var teamObserver: NSObjectProtocol!
     private var conversationListObserver: NSObjectProtocol!
 
-    override init(account: Account) {
+    override init(account: Account, user: ZMUser? = nil) {
         
         imageView = TeamImageView(account: account)
         
-        super.init(account: account)
+        super.init(account: account, user: user)
         
         isAccessibilityElement = true
         accessibilityTraits = UIAccessibilityTraitButton
